Fix document viewer UI and Cloudinary raw file extension issue

+++ BE/src/main/java/com/hourlink/helprequest/dto/response/HelpRequestResponse.java
@@ -19,7 +19,7 @@ public class HelpRequestResponse {
     String desiredTime;
     Integer duration;
     String region;
-    Integer timeCreditAmount;
+    Double timeCreditAmount;
     RequestStatus status;
     UUID categoryId;
     String categoryName;
